I hate cmake, I dont want work on c++ anymore

+++ lab_sound_processor/sound_processor/libs/converters/mix_converter/mix_converter.cpp
@@ -1,4 +1,5 @@
-#include "converters.h"
+#include "converters.hpp"
+//#include
 /*
 params{taskFinished(0), secondsStart(1), secondsEnd(2), changedSeconds(3)}
 sample1 = FinalStream
@@ -7,7 +8,7 @@ sample2 = array of inputSamples
 void conv::MixConverter::process(sampleBuffer &sample1, std::vector<sampleBuffer> &samples,
                            std::vector<size_t> &params) {
   if (params[1] <= params[3] && params[3] < params[2]) {
-    for( int i = 0; i < sample1.curLen_; ++i ) {
+    for( size_t i = 0; i < sample1.curLen_; ++i ) {
       sample1.sample_[i] = sample1.sample_[i] / 2 + samples[0].sample_[i] / 2;
     }
   }
